HDF5 Utilities

To use HDF5 functionality,

pkg> add HDF5

julia> using Quaycle
julia> using HDF5

This package provides a bunch of utilities, such as storing simulation properties and writing ODE solution on the fly to HDF5 format file for postprocessing on different platform and package version. We currently don't support automatically saving green's function to HDF5 due to its limitation of natively storing complex number.

It's worth mention that JLD and JLD2 are also excellent alternatives but they storing the whole type information that may broke read when this package changes or remove certain type definitions. Users may choose them as auxiliary tools.

Public Interface

Quaycle.store — Method

store(filename::AbstractString, p::AbstractProperty)

Store property in HDF5.

Arguments

filename::AbstractString: file name to be used

p::AbstractProperty: property to be saved

Quaycle.@getprop — Macro

@getprop filename

Read property stored in HDF5.

Arguments

filename: file name. Assume it has one and only one kind of property group.

Quaycle.@store — Macro

@store filename::AbstractString p::AbstractProperty

Macro shortcut for storing property in HDF5.

Arguments

filename::AbstractString: file name to be used

p::AbstractProperty: property to be saved