Friction

This package adopts what is called rate-and-state friction as one of the essential components for modeling fault dynamics. Currently, we only support single state variable $θ$.

Public Interface

Quaycle.CForm — Type

Conventional form, see friction

Quaycle.DieterichStateLaw — Type

\[\frac{\mathrm{d}θ}{\mathrm{d}t} = 1 - \frac{V θ}{L}\]

Quaycle.PrzStateLaw — Type

\[\frac{\mathrm{d}θ}{\mathrm{d}t} = 1 - \left(\frac{V θ}{2L}\right)^2\]

Quaycle.RForm — Type

Regularized form, see friction

Quaycle.RuinaStateLaw — Type

\[\frac{\mathrm{d}θ}{\mathrm{d}t} = -\frac{V θ}{L} * \log{\left(\frac{V θ}{L}\right)}\]

Quaycle.StateEvolutionLaw — Type

Currently support:

DieterichStateLaw

RuinaStateLaw

PrzStateLaw

Quaycle.friction — Method

friction(::CForm, v::T, θ::T, a::T, b::T, L::T, f0::T, v0::T) where T

Calculate friction given by the form of fomula as well as other necessary parameters.

Conventional Form:

\[f(V, θ) = f_0 + a \ln{\left(\frac{V}{V_0}\right)} + b \ln{\left(\frac{V_0 θ}{L}\right)}\]

Regularized Form:

\[f(V, θ) = a \sinh ^{-1}{\left(\frac{V}{2V_0} \exp{\left(\frac{f_0 + b \ln{\left(V_0 θ/L\right)}}{a}\right)}\right)}\]

References

Dieterich, J. (1979). Modeling of rock friction: 1. Experimental results and constitutive equations. Journal of Geophysical Research: Solid Earth, 84(B5), 2161–2168. https://doi.org/10.1029/JB084iB05p02161

Ruina, A. (1983). Slip instability and state variable friction laws. Journal of Geophysical Research: Solid Earth, 88(B12), 10359–10370. https://doi.org/10.1029/JB088iB12p10359

Rubin, A. M., & Ampuero, J.-P. (2005). Earthquake nucleation on (aging) rate and state faults. Journal of Geophysical Research: Solid Earth, 110(B11). https://doi.org/10.1029/2005JB003686