Simulation Properties

We attribute one struct of all necessary properties for each type of simulation objections.

Public Interface

Quaycle.SingleDofRSFProperty — Type

System property for single degree of freedom under rate-state friction.

Fields

a: contrib from velocity

b: contrib from state

L: critical distance

k: spring stiffness

σ: effective normal stress

η: radiation damping

vpl: plate rate

f0 = 0.6: ref. frictional coeff

v0 = 1e-6: ref. velocity

Quaycle.RateStateQuasiDynamicProperty — Type

System property for multiple fault patches under rate-state friction.

Fields

a: contrib from velocity

b: contrib from state

L: critical distance

σ: effective normal stress

η: radiation damping

vpl: plate rate

f0 = 0.6: ref. frictional coeff

v0 = 1e-6: ref. velocity

Quaycle.DislocationCreepProperty — Type

System properties for plastic deformation of dislocation creep. Please refer (Hirth & Kohlstedt, 2003) for concrete units of each factor.

Fields

A: prefactor

n: power law stress exponent

COH: water content[1]

r: water fugacity exponent

α: melting constant

ϕ: melting fraction

Q: activation energy

P: pressure

Ω: activation volume

T: temperature

Quaycle.DiffusionCreepProperty — Type

System properties for plastic deformation of diffusion creep. Please refer (Hirth & Kohlstedt, 2003) for concrete units of each factor.

Fields

A: prefactor

d: grain size

m: grain size exponent

COH: water content[1]

r: water fugacity exponent

α: melting constant

ϕ: melting fraction

Q: activation energy

P: pressure

Ω: activation volume

T: temperature

Quaycle.PeierlsProperty — Type

System properties for plastic deformation of Peierls Mechanisms. Not implemented yet.

Quaycle.CompositePlasticDeformationProperty — Type

Compose all three type of plastic deformation and other strain-related system properties, see (Kohlstedt & Hansen, 2015). Each field is the overall equivalent factor not dependent on stress.

Fields

disl: dislocation creep

n: stress exponent in dislocation creep

diff: diffusion creep

peie: Peierls mechanisms

dϵref: reference strain rate whose length must equal strain components considered

Quaycle.ViscoelasticMaxwellProperty — Type

Composite property for viscoelastic rheology of maxwell representation.

Fields

pe::RateStateQuasiDynamicProperty: elastic rate-and-state system property

pv::CompositePlasticDeformationProperty: composite plastic deformation system property

Quaycle.compose — Method

compose(pe::RateStateQuasiDynamicProperty{T}, dϵref::AbstractVector, dϵname::AbstractVector, pvs...) where T

Create maxwell viscoelastic system given both rate-and-state and plastic properties.

Arguments

pe::RateStateQuasiDynamicProperty{T}: elastic rate-and-state system property

dϵref: reference strain rate whose length must equal to and order same as strain components considered

pvs...: different type of plastic deformation system properties but no more than three

Quaycle.composite_factor — Method

composite_factor(pv::PlasticDeformationProperty)

Compute an equivalent factor for levarage recomputing during ODE solving.

Arguments

pv::PlasticDeformationProperty: plastic deformation system property

1Some references may refer it as water fugacity $f_{\mathrm{H_{2} O}}$, which is misleading. Since fugacity has the same dimension of chemical potential (Pa) while water content is dimensionless.