Visualize

This package provides some utilities for writing results to VTK (vts, vti, vtu, vtm, pvd, etc...) file for postprocessing with Paraview. Data cache for writing results are also implemented where users could explore other meshing tools.

One short note is that if you store 6 components vector, for example $σ$, Paraview will interpret them in the order of $σ_{xx}$, $σ_{yy}$, $σ_{zz}$, $σ_{xy}$, $σ_{yz}$, $σ_{xz}$. Whereas in our assemble for stress, the order, same as e.g OpenFOAM, is $σ_{xx}$, $σ_{xy}$, $σ_{xz}$, $σ_{yy}$, $σ_{yz}$, $σ_{zz}$.

Notice that these function are not fully tested. If you encounter any problem, please file an issue with your MWE.

Public Interface

Quaycle.vtk_output — Method

vtk_output(f, t::AbstractVector, u::AbstractVector{<:AbstractVecOrMat},
    ustr::AbstractVector{<:AbstractString}, cache::ParaviewOutputCache)

Write time-series results to single-block paraview collection file.

Arguments

f: output file name

t::AbstractVector: time stamp vector

u::AbstractVector{<:AbstractArray}, list of results to be written

ustr::AbstractVector{<:AbstractString}: list results names to be assigned

cache: cache of data conversion, cell information and nodes information

Quaycle.vtk_output — Method

vtk_output(f, u::AbstractVector{<:AbstractVecOrMat},
    ustr::AbstractVector{<:AbstractString}, cache::ParaviewOutputCache)

Write results to single-block VTU file.

Arguments

f: output file name

u::AbstractVector, list of results to be written

ustr::AbstractVector{<:AbstractString}: list results names to be assigned

cache: cache of data conversion, cell information and nodes information

Quaycle.vtk_output — Method

vtk_output(f, u, ustr, cache::AbstractVector{<:ParaviewOutputCache})

Write results to multiple-block VTM file.

Arguments

f: output file name

u, list of block to be written, each contains a list of results in that block

ustr: list of block to be assigned, each contains a list of results names to data in that block

cache: list cache corresponding to each block

Quaycle.vtk_output — Method

vtk_output(f, t, u, ustr, cache::AbstractVector{<:ParaviewOutputCache})

Write results to multiple-block paraview collection file.

Arguments

f: output file name

t::AbstractVector: time stamp vector

u, list of different block data to be written, each contains a list of results in that block

ustr: list of block names to be assigned, each contains a list of results names to data in that block

cache: list cache corresponding to each block

Quaycle.vtk_output — Method

vtk_output(f, p::AbstractProperty, cache::ParaviewOutputCache)

Store simulation property, including all its fields, to VTU file.

Arguments

f: output file name

p: property struct

cache: cache of data conversion, cell information and nodes information

Notice

Not all AbstractProperty are supported. Only those with well defined fieldnames are. For multi-block domain property, save them separately.