Private

Private Interface

Index

JuEQ.HomogeneousElasticProperties

JuEQ.TmpVariable

JuEQ.applied_unit_dislocation

JuEQ.shear_traction

JuEQ.stiffness_periodic_boundary_condition

Interfaces

JuEQ.HomogeneousElasticProperties — Type.

Okada's dc3d only applies on isotropic materials, therefore, elastic modulus are constrained to be scalars.

source

JuEQ.TmpVariable — Type.

Temporal variable in solving ODEs aimed to avoid allocation overheads.

source

JuEQ.applied_unit_dislocation — Method.

For noraml fault, it should of course be [0., -1., 0.]. However, in term of force balance, it is quivalent to thrust fault if dip angle are constrained within [0, π/2] in fact.

The unit of unit dislocation below is the same of v * t at set by user so to avoid normalization step.

JuEQ.shear_traction — Method.

shear_traction(::Type{<:PlaneFault}, u, λ, μ, dip)

Calculate the shear traction on the fault plane w.r.t. fault types.

Arguments

u::AbstractArray{<:Number, 1}: the output from dc3d_okada

λ::Number: Lamé's first parameter

μ::Number: shear modulus

dip::Number: plane dip angle

Reference

A good reference is at Displacement & Strain & Stress.

JuEQ.stiffness_periodic_boundary_condition — Method.

Periodic boundary condition for 2D faults.

Arguments

same as dc3d_okada, see dc3d for details.

ax::T: along-strike fault length

nrept::Integer: (half) number of repetition, as denoted by -npret: nrept

lrept::T: length of repetition interval, see Note below

Note

The buffer block is evenly distributed on the two along-strike edges, each of which contains half of that.