Private

Private Interface

Index

JuEQ.HomogeneousElasticProperties

JuEQ.TmpRSFVariable

JuEQ.applied_unit_dislocation

JuEQ.shear_traction

JuEQ.stiffness_periodic_boundary_condition

JuEQ.stiffness_tensor

Interfaces

JuEQ.HomogeneousElasticProperties — Type.

Okada's dc3d only applies on isotropic materials, therefore, elastic modulus are constrained to be scalars.

source

JuEQ.TmpRSFVariable — Type.

Temporal variable in solving ODEs aimed to avoid allocation overheads.

source

JuEQ.applied_unit_dislocation — Method.

For noraml fault, it should of course be [0., -1., 0.]. However, in term of force balance, it is quivalent to thrust fault if dip angle are constrained within [0, π/2] in fact.

The unit of unit dislocation below is the same of v * t at set by user so to avoid normalization step.

source

JuEQ.shear_traction — Method.

shear_traction(::Type{<:PlaneFault}, u, λ, μ, dip)

Calculate the shear traction on the fault plane w.r.t. fault types.

Arguments

u::AbstractArray{<:Number, 1}: the output from dc3d_okada

λ::Number: Lamé's first parameter

μ::Number: shear modulus

dip::Number: plane dip angle

Reference

A good reference is at Displacement & Strain & Stress.

source

JuEQ.stiffness_periodic_boundary_condition — Method.

Periodic boundary condition for 2D faults.

Arguments

same as dc3d_okada, see dc3d for details.

ax::AbstractVector: along-strike fault length

nrept::Integer: (half) number of repetition, as denoted by -npret: nrept

lrept::Number: length of repetition interval, see Note below

Note

The buffer block is evenly distributed on the two along-strike edges, each of which contains half of that.

source

JuEQ.stiffness_tensor — Method.

stiffness_tensor(fa::PlaneFaultDomain, gd::BoundaryElementGrid, ep::HomogeneousElasticProperties)

Calculate the reduced stiffness tensor. For 2D fault, the final result will be dimensionally reduced to a 3D array due to the translational & reflective & perodic symmetry, such that the tensor contraction will be equivalent to convolution, hence we could use FFT for better performace.

Note

Faults are originated from surface and extends downwards, thus dep = 0

source